Move S3Node to examples
