+++ package.json
@@ -45,6 +45,7 @@
   },
   "devDependencies": {
     "@aws-sdk/client-s3": "^3.18.0",
+    "@aws-sdk/lib-storage": "^3.18.0",
     "@dsnp/test-generators": "^0.0.0-984e60",
     "@typechain/ethers-v5": "^7.0.0",
     "@types/jest": "26.0.22",
@@ -66,6 +67,7 @@
     "lodash": "^4.17.21",
     "npm-run-all": "4.1.5",
     "prettier": "2.2.1",
+    "stream": "^0.0.2",
     "ts-jest": "26.5.4",
     "ts-node": "9.1.1",
     "typechain": "^5.0.0",
@@ -106,7 +108,6 @@
     "titleTemplate": "{title}"
   },
   "dependencies": {
-    "@aws-sdk/lib-storage": "^3.18.0",
     "@dsnp/contracts": "0.1.0",
     "@dsnp/parquetjs": "0.9.0",
     "@dsnp/test-generators": "^0.1.0",
@@ -114,7 +115,6 @@
     "ethers": "^5.3.0",
     "js-sha3": "^0.8.0",
     "request": "^2.88.2",
-    "stream": "^0.0.2",
     "ws": "^7.4.6"
   },
   "peerDependencies": {
