+++ tsconfig.json
@@ -34,7 +34,7 @@
    * See https://www.typescriptlang.org/tsconfig#lib
    */
   "lib": ["es2017", "esnext"],
-  "exclude": ["node_modules/**", "./*.ts", "./src/test/importTest/**"],
+  "exclude": ["node_modules/**", "./*.ts", "./src/test/importTest/**", "./examples/**"],
   "compileOnSave": false,
   "typedocOptions": {
     "readme": "DOCS.md",
